backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
python-multipart==0.0.20
pydantic==2.10.4
python-dotenv==1.0.1
cors==1.0.1
